requirements.txt
absl-py==2.2.2
asttokens==3.0.0
box2d-py==2.3.5
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
etils==1.12.2
exceptiongroup==1.3.0
executing==2.2.0
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.1
glfw==2.9.0
gymnasium==1.1.1
imageio==2.37.0
importlib_resources==6.5.2
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
mujoco==3.3.2
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pygame==2.6.1
Pygments==2.19.1
PyOpenGL==3.1.9
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
six==1.17.0
stack-data==0.6.3
swig==4.3.1
sympy==1.14.0
tk==0.1.0
torch==2.7.0
traitlets==5.14.3
triton==3.3.0
typing_extensions==4.13.2
wcwidth==0.2.13
zipp==3.22.0
